Fixing issues with non-latin characters and trackNum == 0
Hi @dem214!

Thank you for your library. It's really useful for the project I'm currently working on.
I have faced with a couple of issues.

The first one appears when a path to an XSPF file stored locally contains non-Latin (e.g. Cyrillic) characters. The `urify` static method of the `_Parser` class raises the ValueError, and the parsing process just crashes. I'm not a specialist in encoding formats, and I'm not sure if this is the real solution or just a workaround to bypass the problem, but the function that quotes these characters which I introduced lets the correct local paths of this kind be read/parsed at least.

And the second issue is about track numbers (`trackNum`) with zero values. Some popular players (e.g., AIMP) export XSPF files with them, and then load\read them flawlessly. However, in your current version the `trackNum.setter` method raises ValueError for these playlists, and the parsing crashes as well. Modifying the condition, triggering this error, from `value <= 0` to 'value < 0' made it work for me.

I would be grateful for your feedback/suggestions/improvements.

Thanks again!